Create generic function for adding a comment and use it for project comments
